working program, needs polishing
the bot works, although it has a lookahead move tree depth of only 4, as I was trying to keep as closely to a true A* search as possiblewithout coding out every blocking position on the board. only about 10 punishing states for blocking moves. Bot beats about 50% of games and loops the other half for reasons I am unsure of. This is certainly sufficient enough for building a bot based off of A*. will comment and polish code in next commit.